requirements.txt
APScheduler==3.6.3
black==22.3.0
cachetools==4.2.2
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
idna==3.3
lxml==4.9.0
mypy-extensions==0.4.3
numpy==1.22.4
pandas==1.4.2
pandas-datareader==0.10.0
pathspec==0.9.0
platformdirs==2.5.2
python-dateutil==2.8.2
python-dotenv==0.20.0
python-telegram-bot==13.12
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.0
six==1.16.0
tomli==2.0.1
tornado==6.1
typing_extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
